rm conversion tools script
